--- a/pkg/logcli/output/loki.go
+++ b/pkg/logcli/output/loki.go
@@ -1,0 +1,1 @@
+package output