--- a/pkg/pattern/drain/drain.go
+++ b/pkg/pattern/drain/drain.go
@@ -251,7 +251,7 @@ func (d *Drain) TrainPattern(content string, samples []*logproto.PatternSample) 
 	return matchCluster
 }
 
-func deduplicatePlaceholders(line string) string {
+func deduplicatePlaceholders(line string, placeholder string) string {
 	first := strings.Index(line, "<_><_>")
 	if first == -1 {
 		return line
@@ -259,14 +259,14 @@ func deduplicatePlaceholders(line string) string {
 	builder := make([]byte, 0, len(line))
 	low := 0
 	for i := first; i < len(line)-5; i++ {
-		if line[i:i+3] == "<_>" {
+		if line[i:i+len(placeholder)] == placeholder {
 			high := i + 3
 			for ; high < len(line)-2; high += 3 {
-				if line[high:high+3] != "<_>" {
+				if line[high:high+len(placeholder)] != placeholder {
 					break
 				}
 			}
-			builder = append(builder, line[low:i+3]...)
+			builder = append(builder, line[low:i+len(placeholder)]...)
 			low = high
 			i = high
 		}
@@ -277,7 +277,7 @@ func deduplicatePlaceholders(line string) string {
 }
 
 func (d *Drain) PatternString(c *LogCluster) string {
-	s := deduplicatePlaceholders(d.tokenizer.Join(c.Tokens, c.TokenState))
+	s := deduplicatePlaceholders(d.tokenizer.Join(c.Tokens, c.TokenState), d.config.ParamString)
 	if s == d.config.ParamString {
 		return ""
 	}