--- a/pkg/pattern/drain/line_tokenizer.go
+++ b/pkg/pattern/drain/line_tokenizer.go
@@ -1,14 +1,18 @@
 package drain
 
 import (
+	"regexp"
 	"strings"
 	"unicode"
+	"unsafe"
 )
 
 type LineTokenizer interface {
 	Tokenize(line string) ([]string, interface{})
 	Join(tokens []string, state interface{}) string
 }
+
+var pat = regexp.MustCompile("(<_>)+")
 
 type spacesTokenizer struct{}
 
@@ -71,6 +75,31 @@ func (p *punctuationTokenizer) Tokenize(line string) ([]string, interface{}) {
 	}
 
 	return tokens[:nextTokenIdx], spacesAfter[:nextSpaceIdx]
+}
+
+func deduplicatePlaceholders(line string) string {
+	first := strings.Index(line, "<_><_>")
+	if first == -1 {
+		return line
+	}
+	builder := make([]byte, 0, len(line))
+	low := 0
+	for i := first; i < len(line)-5; i++ {
+		if line[i:i+3] == "<_>" {
+			high := i + 3
+			for ; high < len(line)-2; high += 3 {
+				if line[high:high+3] != "<_>" {
+					break
+				}
+			}
+			builder = append(builder, line[low:i+3]...)
+			low = high
+			i = high
+		}
+	}
+	builder = append(builder, line[low:]...)
+
+	return unsafe.String(unsafe.SliceData(builder), len(builder))
 }
 
 func (p *punctuationTokenizer) Join(tokens []string, state interface{}) string {