--- a/pkg/pattern/drain/line_tokenizer_test.go
+++ b/pkg/pattern/drain/line_tokenizer_test.go
@@ -1,8 +1,6 @@
 package drain
 
 import (
-	"fmt"
-	"strings"
 	"testing"
 
 	"github.com/stretchr/testify/require"
@@ -98,62 +96,6 @@ var testCases = []TestCase{
 			typeSplitting:   {`!@£$%^&*()`},
 		},
 	},
-}
-
-func TestDeduplicatePlaceholders(b *testing.T) {
-	type DedupCase struct {
-		line string
-		want string
-	}
-	cases := []DedupCase{
-		{
-			line: "abcd",
-			want: "abcd",
-		},
-		{
-			line: "<_><_>abcd",
-			want: "<_>abcd",
-		},
-		{
-			line: strings.Repeat("<_>", 100),
-			want: "<_>",
-		},
-		{
-			line: "<_> <_>",
-			want: "<_> <_>",
-		},
-		{
-			line: strings.Repeat("<_> ", 100),
-			want: strings.Repeat("<_> ", 100),
-		},
-		{
-			line: "<_><<_>",
-			want: "<_><<_>",
-		},
-		{
-			line: "<_><->",
-			want: "<_><->",
-		},
-		{
-			line: strings.Repeat(strings.Repeat("<_>", 100)+" ", 100),
-			want: strings.Repeat("<_> ", 100),
-		},
-		{
-			line: "<<<<<<<_><_>>>>>>>>",
-			want: "<<<<<<<_>>>>>>>>",
-		},
-		{
-			line: strings.Repeat("A", 100) + "<_><_>",
-			want: strings.Repeat("A", 100) + "<_>",
-		},
-	}
-
-	for i, tc := range cases {
-		b.Run(fmt.Sprintf("Dedup %d", i), func(t *testing.T) {
-			got := deduplicatePlaceholders(tc.line)
-			require.Equal(t, tc.want, got)
-		})
-	}
 }
 
 func TestTokenizer_Tokenize(t *testing.T) {