--- a/pkg/pattern/drain/log_cluster.go
+++ b/pkg/pattern/drain/log_cluster.go
@@ -22,9 +22,9 @@ type LogCluster struct {
 
 func (c *LogCluster) String() string {
 	if c.Stringer != nil {
-		return c.Stringer(c.Tokens, c.TokenState)
+		return deduplicatePlaceholders(c.Stringer(c.Tokens, c.TokenState))
 	}
-	return strings.Join(c.Tokens, " ")
+	return deduplicatePlaceholders(strings.Join(c.Tokens, " "))
 }
 
 func (c *LogCluster) append(ts model.Time) {