… pkg/pattern/drain/testdata/istio-json.txt
… pkg/pattern/ingester_querier_test.go